tests/wpt/tests/html/browsers/history/the-location-interface/location_port.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_port</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      var port = location.port;
      var url = location.href;

      var pos = url.indexOf("//");
      if (pos != -1) {
        url = url.substr(pos+2, url.length-pos-2);
        pos = url.indexOf("/");
        if (pos != -1)
          url = url.substr(0, pos);
        pos = url.indexOf(":");
        if (pos != -1)
          url = url.substr(pos+1, url.length-pos-1);
      }

      assert_equals(port, url, "port");

    }, "location port");
    </script>
  </body>
</html>
